+++ packages/plugins/notifier-slack/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @composio/ao-plugin-notifier-slack
+# @agentboard/ao-plugin-notifier-slack
 
 ## 0.2.0
 
 ### Patch Changes
 
 - Updated dependencies [3a650b0]
-  - @composio/ao-core@0.2.0
+  - @agentboard/ao-core@0.2.0